BAUDRATE from int to long.

+++ examples/TestConnection/TestConnection.ino
@@ -3,7 +3,7 @@
 #include "Streaming.h"
 #include "TMC26X.h"
 
-const int BAUDRATE = 115200;
+const long BAUDRATE = 115200;
 const int LOOP_DELAY = 2000;
 const int CS_PIN = 9;
